+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5462610_sys_gh1489-disable-cockpit-reference-in-order-window.sql
@@ -1,0 +1,15 @@
+-- 2017-05-15T17:44:14.624
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Window SET IsActive='N',Updated=TO_TIMESTAMP('2017-05-15 17:44:14','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Window_ID=540018
+;
+
+-- 2017-05-15T17:44:14.626
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Menu SET Description=NULL, IsActive='N', Name='Cockpit',Updated=TO_TIMESTAMP('2017-05-15 17:44:14','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Menu_ID=540029
+;
+
+-- 2017-05-15T17:44:37.036
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Table SET IsActive='N',Updated=TO_TIMESTAMP('2017-05-15 17:44:37','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Table_ID=540060
+;
+